requirements.txt
pymongo==3.9.0
flask==1.1.1
Flask-JWT==0.3.2